Make extract link functionality a utility
