--- Dockerfile.integration
@@ -1,3 +1,0 @@
-FROM node:18-alpine
-
-RUN apk add rsync
