+++ docker-compose-dev.yml
@@ -5,8 +5,6 @@ services:
     volumes: [ node-modules:/work/node_modules ]
   script:
     volumes: [ node-modules:/work/node_modules ]
-  integration:
-    volumes: [ node-modules:/work/node_modules ]
   legacy:
     volumes: [ bundle:/usr/local/bundle ]
   ruby-script:
