+++ src/utils/__tests__/examples/duplicate-paths/api.html
@@ -1,0 +1,5 @@
+<!DOCTYPE html>
+<html lang="en-US">
+  <a href="#heading-1">
+  <a href="about:blank#heading-2">
+</html>
