+++ src/utils/__tests__/examples/duplicate-paths/api/index.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html lang="en-US">
+  <a href="#heading-1">
+  <a href="#heading-1">
+  <a href="#heading-2">
+  <a href="#heading-3">
+  <a href="#heading-4">
+  <a href="#heading-5">
+</html>
