+++ src/utils/__tests__/examples/empty-anchor/api.html
@@ -1,0 +1,4 @@
+<!DOCTYPE html>
+<html lang="en-US">
+  <a href="#">
+</html>
