+++ src/utils/__tests__/examples/external-links/api.html
@@ -1,0 +1,5 @@
+<!DOCTYPE html>
+<html lang="en-US">
+  <a href="#heading-1">
+  <a href="https://www.external.com">
+</html>
